Streamlined the MenuItems and PopulateMenus routes
